renames, let's do all folders lowercase
