+++ src/components/plots/FixedTicks.tsx
