+++ src/components/plots/PointCloud.tsx
