+++ src/components/plots/UVCube.tsx
