+++ src/components/plots/index.ts
@@ -2,5 +2,5 @@ export {DataCube} from './DataCube';
 export {UVCube} from './UVCube';
 export {PlotLine} from './PlotLine'
 export {PointCloud} from './PointCloud'
-export {PlotArea} from './PlotArea/PlotArea'
+export {PlotArea} from './plotarea/PlotArea'
 export {FixedTicks} from './FixedTicks'
