+++ src/components/plots/plotarea/PlotArea.css
